Add Autotemp binary + refactor binary
