+++ custom_components/ithodaalderop/sensors/base.py
@@ -1,7 +1,12 @@
 """Sensor base class."""
 
+import json
+
+from homeassistant.components import mqtt
+from homeassistant.components.binary_sensor import BinarySensorEntity
 from homeassistant.components.sensor import SensorEntity
 from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import callback
 from homeassistant.helpers.device_registry import DeviceInfo
 
 from ..const import (
@@ -13,7 +18,10 @@ from ..const import (
     NONCVE_DEVICES,
     UNITTYPE_ICONS,
 )
-from ..definitions.base import IthoSensorEntityDescription
+from ..definitions.base import (
+    IthoBinarySensorEntityDescription,
+    IthoSensorEntityDescription,
+)
 
 
 class IthoBaseSensor(SensorEntity):
@@ -68,3 +76,69 @@ class IthoBaseSensor(SensorEntity):
         if self.entity_description.native_unit_of_measurement in UNITTYPE_ICONS:
             return UNITTYPE_ICONS[self.entity_description.native_unit_of_measurement]
         return None
+
+
+class IthoBinarySensor(BinarySensorEntity):
+    """Representation of a Itho add-on binary sensor that is updated via MQTT."""
+
+    _attr_has_entity_name = True
+    entity_description: IthoBinarySensorEntityDescription
+
+    def __init__(
+        self,
+        description: IthoBinarySensorEntityDescription,
+        config_entry: ConfigEntry,
+    ) -> None:
+        """Initialize the binary sensor."""
+        self.entity_description = description
+        self.entity_description.translation_key = self.entity_description.key
+
+        model = ADDON_TYPES[config_entry.data[CONF_ADDON_TYPE]]
+        if config_entry.data[CONF_ADDON_TYPE] == "noncve":
+            model = f"{model} - {NONCVE_DEVICES[config_entry.data[CONF_NONCVE_MODEL]]}"
+
+        self._attr_device_info = DeviceInfo(
+            identifiers={(DOMAIN, config_entry.data[CONF_ADDON_TYPE])},
+            manufacturer=MANUFACTURER,
+            model=model,
+            name=f"Itho Daalderop {ADDON_TYPES[config_entry.data[CONF_ADDON_TYPE]]}",
+        )
+
+        self._attr_unique_id = (
+            f"itho_{ADDON_TYPES[config_entry.data[CONF_ADDON_TYPE]]}_{description.key}"
+        )
+        self.entity_id = f"binary_sensor.{self._attr_unique_id}"
+
+    @property
+    def icon(self):
+        """Icon for binary sensor."""
+        if (
+            self.entity_description.icon_off is not None
+            and self.entity_description.icon_on is not None
+        ):
+            if self._attr_is_on:
+                return self.entity_description.icon_on
+            return self.entity_description.icon_off
+        if self.entity_description.icon is not None:
+            return self.entity_description.icon
+        return None
+
+    async def async_added_to_hass(self) -> None:
+        """Subscribe to MQTT events."""
+
+        @callback
+        def message_received(message):
+            """Handle new MQTT messages."""
+            payload = json.loads(message.payload)
+            json_field = self.entity_description.json_field
+            if json_field not in payload:
+                value = None
+            else:
+                value = bool(payload[json_field])
+
+            self._attr_is_on = value
+            self.async_write_ha_state()
+
+        await mqtt.async_subscribe(
+            self.hass, self.entity_description.topic, message_received, 1
+        )
